+++ src/hic/HIGenHeader.cpp
@@ -48,13 +48,13 @@ bool HIGenHeader::run()
              "#define HIC_" << idstr << "\n\n"
              "class QObject;\n"
              "\n"
-             "#include \"libHeaven/Actions/Action.h\"\n"
-             "#include \"libHeaven/Actions/WidgetAction.h\"\n"
-             "#include \"libHeaven/Actions/Menu.h\"\n"
-             "#include \"libHeaven/Actions/MenuBar.h\"\n"
-             "#include \"libHeaven/Actions/ToolBar.h\"\n"
-             "#include \"libHeaven/Actions/MergePlace.h\"\n"
-             "#include \"libHeaven/Actions/ActionContainer.h\"\n"
+             "#include \"libHeaven/Actions/Action.hpp\"\n"
+             "#include \"libHeaven/Actions/WidgetAction.hpp\"\n"
+             "#include \"libHeaven/Actions/Menu.hpp\"\n"
+             "#include \"libHeaven/Actions/MenuBar.hpp\"\n"
+             "#include \"libHeaven/Actions/ToolBar.hpp\"\n"
+             "#include \"libHeaven/Actions/MergePlace.hpp\"\n"
+             "#include \"libHeaven/Actions/ActionContainer.hpp\"\n"
              "#include \"libHeaven/Actions/DynamicActionMerger.hpp\"\n"
              "\n";
