+++ src/libHeaven/Actions/Menu.cpp
@@ -17,15 +17,15 @@
 #include <QMenu>
 #include <QAction>
 
-#include "libHeaven/Actions/MenuPrivate.h"
-#include "libHeaven/Actions/ActionPrivate.h"
+#include "libHeaven/Actions/MenuPrivate.hpp"
+#include "libHeaven/Actions/ActionPrivate.hpp"
 #include "libHeaven/Actions/DynamicActionMerger.hpp"
 #include "libHeaven/Actions/DynamicActionMergerPrivate.hpp"
-#include "libHeaven/Actions/ActionContainerPrivate.h"
-#include "libHeaven/Actions/MergePlacePrivate.h"
-#include "libHeaven/Actions/Separator.h"
-#include "libHeaven/Actions/UiManager.h"
-#include "libHeaven/Actions/MergesManager.h"
+#include "libHeaven/Actions/ActionContainerPrivate.hpp"
+#include "libHeaven/Actions/MergePlacePrivate.hpp"
+#include "libHeaven/Actions/Separator.hpp"
+#include "libHeaven/Actions/UiManager.hpp"
+#include "libHeaven/Actions/MergesManager.hpp"
 
 namespace Heaven
 {
