+++ src/libHeaven/Actions/UiContainer.cpp
@@ -18,16 +18,16 @@
 #include <QMenuBar>
 #include <QToolBar>
 
-#include "libHeaven/Actions/UiContainer.h"
-#include "libHeaven/Actions/UiManager.h"
-#include "libHeaven/Actions/ActionContainerPrivate.h"
-#include "libHeaven/Actions/MenuPrivate.h"
-#include "libHeaven/Actions/ActionPrivate.h"
-#include "libHeaven/Actions/WidgetActionPrivate.h"
-#include "libHeaven/Actions/MergePlacePrivate.h"
-#include "libHeaven/Actions/Separator.h"
-#include "libHeaven/Actions/MergesManager.h"
-#include "libHeaven/Actions/WidgetActionWrapper.h"
+#include "libHeaven/Actions/UiContainer.hpp"
+#include "libHeaven/Actions/UiManager.hpp"
+#include "libHeaven/Actions/ActionContainerPrivate.hpp"
+#include "libHeaven/Actions/MenuPrivate.hpp"
+#include "libHeaven/Actions/ActionPrivate.hpp"
+#include "libHeaven/Actions/WidgetActionPrivate.hpp"
+#include "libHeaven/Actions/MergePlacePrivate.hpp"
+#include "libHeaven/Actions/Separator.hpp"
+#include "libHeaven/Actions/MergesManager.hpp"
+#include "libHeaven/Actions/WidgetActionWrapper.hpp"
 
 namespace Heaven
 {
