+++ src/libHeaven/App/PrimaryWindow.cpp
@@ -28,7 +28,7 @@
 #include "libHeaven/App/PrimaryWindow.hpp"
 #include "libHeaven/App/PrimaryWindowPrivate.hpp"
 
-#include "libHeaven/Actions/MenuBar.h"
+#include "libHeaven/Actions/MenuBar.hpp"
 
 #include "libHeaven/Widgets/ModeSwitchWidget.hpp"
 #include "libHeaven/Widgets/FooterWidget.hpp"
