+++ src/libHeaven/MultiBar/MultiBarToolSection.cpp
@@ -17,7 +17,7 @@
 #include <QToolBar>
 #include <QVariant>
 
-#include "Actions/ToolBar.h"
+#include "Actions/ToolBar.hpp"
 
 #include "MultiBar/MultiBarToolSection.hpp"
 #include "MultiBar/MultiBarSectionPrivate.hpp"
